parte do motorista 60% finalizada

+++ src/models/Vans.js
@@ -4,7 +4,7 @@ class Vans extends Model {
     static init(connection) {
         super.init({
             placa: DataTypes.STRING,
-            renavam: DataTypes.INTEGER,
+            renavam: DataTypes.BIGINT,
             modelo: DataTypes.STRING,
             marca: DataTypes.STRING,
             ano: DataTypes.INTEGER,
